requirements.txt
altair==5.5.0
annotated-types==0.7.0
attrs==25.3.0
blinker==1.9.0
cachelib==0.13.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cmake==3.31.4
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
Cython==3.0.12
dash==3.0.1
dash-bootstrap-components==2.0.0
dash-extensions==2.0.0
dataclass-wizard==0.35.0
EditorConfig==0.17.0
filelock==3.17.0
Flask==3.0.3
Flask-Caching==2.3.1
fonttools==4.56.0
fsspec==2025.3.0
geopandas==1.0.1
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.1.1
idna==3.10
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.6
jsbeautifier==1.15.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
lottie==0.7.1
MarkupSafe==3.0.2
matplotlib==3.10.1
more-itertools==10.6.0
mpmath==1.3.0
narwhals==1.32.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.2.3
-e c:\users\siu856557706\nuplan-devkit
packaging==24.2
pandas==2.2.3
pillow==11.1.0
plotly==6.0.1
protobuf==5.29.4
psycopg2-binary==2.9.10
pyarrow==19.0.1
pydantic==2.11.1
pydantic_core==2.33.0
pydeck==0.9.1
pyogrio==0.10.0
pyparsing==3.2.1
pyproj==3.7.1
pyquaternion==0.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.1
referencing==0.36.2
requests==2.32.3
retrying==1.3.4
rpds-py==0.24.0
scipy==1.15.2
setuptools==76.0.0
shapely==2.0.7
six==1.17.0
smmap==5.0.2
SQLAlchemy==1.4.39
streamlit==1.44.1
sympy==1.13.1
tabulate==0.9.0
tenacity==9.1.2
toml==0.10.2
torch==2.6.0
torchvision==0.21.0
tornado==6.4.2
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
watchdog==6.0.0
Werkzeug==3.0.6
wheel==0.45.1
zipp==3.21.0
